// src/modules/dadata-api/interfaces/dadata-api.enum.ts
export enum DadataEmailType {
    PERSONAL = 'PERSONAL',
    CORPORATE = 'CORPORATE',
    ROLE = 'ROLE',
    DISPOSABLE = 'DISPOSABLE',
}

export enum DadataBankType {
    BANK = 'BANK',
    BANK_BRANCH = 'BANK_BRANCH',
    NKO = 'NKO',
    NKO_BRANCH = 'NKO_BRANCH',
    RKC = 'RKC',
    CBR = 'CBR',
    TREASURY = 'TREASURY',
    OTHER = 'OTHER',
}

export enum DadataTypes {
    ADDRESS = 'address',
    COMPANY = 'party',
    BANK = 'bank',
    EMAIL = 'email',
    FMSUNIT = 'fms_unit',
    FIO = 'fio',
    PASSPORT = 'passport',
    NAME = 'name',
    PHONE = 'phone',
    //GEOLOCATEADDRESS = 'address'
}
export enum SuggestionsStatus {
    YES = 'YES',
    NO = 'NO',
}